requirements.txt
cmake (0.4.0)
colorama (0.3.7)
constantly (15.1.0)
cssselect (1.0.0)
decorator (4.0.10)
Django (1.10.3)
django-admin-tools (0.8.0)
django-leaflet (0.20.0.dev0, c:\users\user\src\django-leaflet)
django-redis-cache (1.7.1)
GDAL (2.1.2)
incremental (16.10.1)
ipdb (0.10.1)
ipython (5.1.0)
ipython-genutils (0.1.0)
jsonfield (1.0.3)
lxml (3.6.4)
mysqlclient (1.3.8)
path.py (9.0)
pexpect (4.2.1)
pickleshare (0.7.4)
Pillow (3.4.2)
pip (9.0.1)
prompt-toolkit (1.0.9)
psycopg2 (2.6.2)
ptyprocess (0.5.1)
pycairo (1.10.0)
Pygments (2.1.3)
pyproj (1.9.5.1)
pytz (2016.7)
redis (2.10.5)
reportlab (3.3.0)
setuptools (20.10.1)
shell (1.0.1)
simplegeneric (0.8.1)
six (1.10.0)
traitlets (4.3.1)
Twisted (16.5.0)
wcwidth (0.1.7)
wheel (0.29.0)
win-unicode-console (0.5)
zope.interface (4.3.2)